backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-socketio==5.11.0
python-multipart==0.0.9
supabase==2.7.4
google-generativeai==0.8.0
elevenlabs==1.8.0
requests==2.32.3
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
redis==5.1.1
aiohttp==3.10.0
pydantic==2.9.0
pydantic-settings==2.5.0
